requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.0
cloudinary==1.30.0
cryptography==38.0.1
Django==4.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12

numpy~=1.23.4